Fix size_t/uint64_t comparison warning on 32-bit targets
Cast totalLen to uint32_t after the UINT32_MAX check passes, avoiding
implicit narrowing comparison between uint64_t and size_t on platforms
where size_t is 32-bit.

+++ src/lib/asn1/ASN1Reader.cpp
@@ -57,7 +57,7 @@ CHIP_ERROR ASN1Reader::Next()
     // could wrap around, potentially advancing mElemStart to an unintended location.
     uint64_t totalLen = static_cast<uint64_t>(mHeadLen) + ValueLen;
     VerifyOrReturnError(totalLen <= UINT32_MAX, ASN1_ERROR_LENGTH_OVERFLOW);
-    VerifyOrReturnError(totalLen <= static_cast<size_t>(mContainerEnd - mElemStart), ASN1_END);
+    VerifyOrReturnError(static_cast<uint32_t>(totalLen) <= static_cast<uint32_t>(mContainerEnd - mElemStart), ASN1_END);
 
     mElemStart = mElemStart + mHeadLen + ValueLen;
 
@@ -145,7 +145,7 @@ CHIP_ERROR ASN1Reader::ExitContainer()
         ResetElementState();
         return ASN1_ERROR_LENGTH_OVERFLOW;
     }
-    if (totalLen > static_cast<size_t>(prevContext.ContainerEnd - prevContext.ElemStart))
+    if (static_cast<uint32_t>(totalLen) > static_cast<uint32_t>(prevContext.ContainerEnd - prevContext.ElemStart))
     {
         ResetElementState();
         return ASN1_END;
